bkp/sic-003-mbpp-tms-api/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
boto==2.49.0
boto3==1.9.242
botocore==1.12.242
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.5
django-allauth==0.40.0
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
drf-extensions==0.5.0
freeze==1.0.10
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.20
jmespath==0.9.4
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
vine==1.3.0
whitelist==0.2.8
whitenoise==4.1.4
zipp==0.6.0
